star commands

+++ discord-bot/commands/deletereview.js
@@ -65,10 +65,12 @@ module.exports = {
         }
 
         // Update user stats
-        const array_user = db.user_stats.get(interaction.user.id, 'review_list');
+        let array_user = db.user_stats.get(interaction.user.id, 'review_list');
         for (let i = 0; i < array_user.length; i++) {
-            // if (array_user[i].includes(songName)) delete array_user[i];
+            if (array_user[i].includes(songName)) array_user = array_user.filter(x => x !== array_user[i]);
         }
+
+        db.user_stats.set(interaction.user.id, array_user, 'review_list');
 
         if (db.user_stats.get(interaction.user.id, 'recent_review').includes(songName)) {
             db.user_stats.set(interaction.user.id, 'N/A', 'recent_review');
@@ -81,16 +83,18 @@ module.exports = {
 
             if (rname === undefined) break;
 
-            //let reviewMsgID;
-
             songObj = db.reviewDB.get(artistArray[i], `["${songName}"]`);
-            //reviewMsgID = db.reviewDB.get(artistArray[i], `["${songName}"].["${userToDelete.id}"].msg_id`);
+            let reviewMsgID = db.reviewDB.get(artistArray[i], `["${songName}"].["${userToDelete.id}"].msg_id`);
+            console.log(reviewMsgID);
+
             delete songObj[`${userToDelete.id}`];
 
-            /*let channelsearch = interaction.guild.channels.cache.get(db.server_settings.get(interaction.guild.id, 'review_channel').slice(0, -2).slice(1));
-            channelsearch.messages.fetch(`${reviewMsgID}`).then(msg => {
-                msg.delete();
-            });*/
+            if (reviewMsgID != false && reviewMsgID != undefined) {
+                let channelsearch = interaction.guild.channels.cache.get(db.server_settings.get(interaction.guild.id, 'review_channel').slice(0, -1).slice(2));
+                channelsearch.messages.fetch(`${reviewMsgID}`).then(msg => {
+                    msg.delete();
+                });
+            }
 
             db.reviewDB.set(artistArray[i], songObj, `["${songName}"]`);
         }
